+++ vimiv/api/_modules.py
@@ -26,6 +26,8 @@ _logger = log.module_logger(__name__)
 ###############################################################################
 #                                  Commands                                   #
 ###############################################################################
+@api.keybindings.register("<button-middle>", "enter thumbnail", mode=api.modes.LIBRARY)
+@api.keybindings.register("<button-middle>", "enter thumbnail", mode=api.modes.IMAGE)
 @api.keybindings.register("<button-right>", "enter library", mode=api.modes.THUMBNAIL)
 @api.keybindings.register("<button-right>", "enter library", mode=api.modes.IMAGE)
 @api.keybindings.register("gm", "enter manipulate")